fix: 로그인 헨들러 추출.
이제 회원가입 헨들러에서 로그인과 회원가입을 처리합니다.
