+++ src/utils/parser/loginhandler.js
@@ -1,0 +1,7 @@
+
+
+const loginHandler = async ({ socket, sequence, payload }) => {
+  
+};
+
+export default loginHandler;
